Разработка модулей #4 Передача температуры по радио каналу.

+++ libraries/SmartNet/src/NetComponent.h
@@ -14,16 +14,16 @@ class NetComponent {
         uint16_t receiver = NULL;
         uint16_t rport = NULL;
         uint16_t timeout = NULL;
-        uint8_t cmd = NULL;
+        uint16_t cmd = NULL;
         unsigned long last = 0;
     };
 
 public:
     NetComponent(uint16_t sp, SmartNet *n);
 
-    void addReceiver(uint16_t r, uint16_t rp, uint8_t c, uint16_t t);
+    void addReceiver(uint16_t r, uint16_t rp, uint16_t c, uint16_t t);
 
-    virtual void sendCommandData(uint16_t r, uint16_t rp, uint8_t cmd) = 0;
+    virtual void sendCommandData(uint16_t r, uint16_t rp, uint16_t cmd) = 0;
 
     void tick();
